Inline graphql-codegen deps
